fix(contracts): consolidate duplicate PayloadStore protocols (QW-03)
Consolidate two separate PayloadStore protocol definitions into single
source of truth in contracts/ layer:

- Create contracts/payload_store.py with full PayloadStore Protocol
- Update core/payload_store.py to import from contracts (re-export)
- Update core/retention/purge.py to import from contracts
- Remove duplicate PayloadStoreProtocol (was "minimal protocol to
  avoid circular imports")

The duplicate existed because purge.py couldn't import from payload_store.py
without circular imports. Moving to contracts/ resolves this architecturally.

From RC1-remediation.md Phase 0 Quick Wins.
